pyproject.toml
[tool.poetry]
name = "verbose-fiesta"
version = "0.0.1"
description = "A simple Flask Hello World application"
authors = ["Leandro E. Colombo Viña <lecovi@zenbook>"]
license = "AGPL-3.0-or-later"

[tool.poetry.dependencies]
python = "^3.7"
flask = "^1.1.2"
python-dotenv = "^0.13.0"
flask-sqlalchemy = "^2.4.4"
psycopg2-binary = "^2.8.6"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
